requirements.txt
Flask
bcrypt
PyJWT
python-dotenv
